Refactor: DetailPage 내부 요소들 컴포넌트화 (Bannerm, ButtonGroup)
