+++ src/components/detail/ButtonGroup/ButtonGroup.module.css
@@ -1,0 +1,9 @@
+.btn-group {
+  display: flex;
+  gap: 0.5rem;
+  justify-content: flex-end;
+}
+
+.btn-group button {
+  width: fit-content;
+}
